refactor: move github package from common to suite
