+++ packages/suite-desktop-ui/package.json
@@ -14,10 +14,10 @@
     },
     "dependencies": {
         "@sentry/electron": "^7.10.0",
-        "@suite-common/github": "workspace:*",
         "@suite/analytics": "workspace:*",
         "@suite/desktop-update": "workspace:*",
         "@suite/flags": "workspace:*",
+        "@suite/github": "workspace:*",
         "@suite/intl": "workspace:*",
         "@suite/router": "workspace:*",
         "@suite/sentry": "workspace:*",
