+++ packages/suite-desktop-ui/tsconfig.json
@@ -6,10 +6,10 @@
     },
     "include": ["."],
     "references": [
-        { "path": "../../suite-common/github" },
         { "path": "../../suite/analytics" },
         { "path": "../../suite/desktop-update" },
         { "path": "../../suite/flags" },
+        { "path": "../../suite/github" },
         { "path": "../../suite/intl" },
         { "path": "../../suite/router" },
         { "path": "../../suite/sentry" },
