+++ packages/suite/package.json
@@ -41,7 +41,6 @@
         "@suite-common/firmware-authenticity": "workspace:*",
         "@suite-common/formatters": "workspace:*",
         "@suite-common/geolocation": "workspace:*",
-        "@suite-common/github": "workspace:*",
         "@suite-common/logger": "workspace:*",
         "@suite-common/message-system": "workspace:*",
         "@suite-common/metadata-types": "workspace:*",
@@ -85,6 +84,7 @@
         "@suite/feature-feedback": "workspace:*",
         "@suite/firmware-upgrade": "workspace:*",
         "@suite/flags": "workspace:*",
+        "@suite/github": "workspace:*",
         "@suite/idb-migration-utils": "workspace:*",
         "@suite/intl": "workspace:*",
         "@suite/locks": "workspace:*",
