+++ packages/suite/src/views/settings/SettingsDebug/GithubIssue.tsx
@@ -1,4 +1,4 @@
-import { openGithubIssue } from '@suite-common/github';
+import { openGithubIssue } from '@suite/github';
 import { ActionButton, ActionColumn, SectionItem, TextColumn } from '@trezor/product-components';
 
 import { useDevice, useSelector } from 'src/hooks/suite';
