+++ suite/github/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@suite-common/github",
+    "name": "@suite/github",
     "version": "1.0.0",
     "private": true,
     "license": "See LICENSE.md in repo root",
